refactor: changes the name from src/modeler to src/unicorn and updates the imports on the files
